- Add camera setting in Render Settings
- Color field send signal when color modified (don't need apply) and implement cancel
- Debug FloatField and IntField

+++ src/ui/widget/custom_widget/float_field_widget.cpp
@@ -42,7 +42,7 @@ namespace VTX::UI::Widget::CustomWidget
 
 	void FloatFieldWidget::_onTextFieldEdited()
 	{
-		const uint newValue = _textField->text().toInt();
+		const float newValue = _textField->text().toFloat();
 		if ( newValue != _value )
 		{
 			setValue( newValue );
